Simplify vite-app template
Move ember-data into the specific test scenario thats wants it, so that the base vite-app template does't need to have ember-data.
