--- tests/vite-app/app/models/post.js
@@ -1,4 +1,0 @@
-import Model, { attr } from '@ember-data/model';
-export default class extends Model {
-  @attr message;
-}
