+++ tests/vite-app/app/routes/application.ts
@@ -1,8 +1,7 @@
 import Route from '@ember/routing/route';
-import { service } from '@ember/service';
+
 export default class extends Route {
-  @service store;
   async model() {
-    return await this.store.findRecord('post', 1);
+    return { message: 'Hello world' };
   }
 }
