+++ tests/vite-app/package.json
@@ -45,20 +45,19 @@
     "ember-auto-import": "^2.6.3",
     "ember-cli": "~5.0.0",
     "ember-cli-app-version": "^6.0.0",
-    "ember-cli-babel": "^7.26.11",
+    "ember-cli-babel": "^8.2.0",
     "ember-cli-clean-css": "^2.0.0",
     "ember-cli-dependency-checker": "^3.3.1",
     "ember-cli-htmlbars": "^6.2.0",
     "ember-cli-inject-live-reload": "^2.1.0",
     "ember-cli-sri": "^2.1.1",
     "ember-cli-terser": "^4.0.2",
-    "ember-data": "~5.3.0",
     "ember-load-initializers": "^2.1.2",
     "ember-modifier": "^4.1.0",
     "ember-page-title": "^7.0.0",
     "ember-qunit": "^7.0.0",
     "ember-resolver": "^10.1.0",
-    "ember-source": "~5.1.0",
+    "ember-source": "5.7.0-beta.1",
     "ember-template-lint": "^5.10.3",
     "ember-welcome-page": "^7.0.2",
     "eslint": "^8.42.0",
@@ -67,8 +66,10 @@
     "eslint-plugin-n": "^16.0.0",
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-qunit": "^7.3.4",
+    "js-reporters": "^2.1.0",
     "loader.js": "^4.7.0",
     "prettier": "^2.8.8",
+    "puppeteer-chromium-resolver": "^21.0.0",
     "qunit": "^2.19.4",
     "qunit-dom": "^2.0.0",
     "stylelint": "^15.7.0",
@@ -76,9 +77,7 @@
     "stylelint-prettier": "^3.0.0",
     "tracked-built-ins": "^3.1.1",
     "typescript": "^5.1.6",
-    "vite": "^5.0.9",
-    "js-reporters": "^2.1.0",
-    "puppeteer-chromium-resolver": "^21.0.0"
+    "vite": "^5.0.9"
   },
   "engines": {
     "node": "16.* || >= 18"
