--- tests/vite-app/public/posts/1.json
@@ -1,9 +1,0 @@
-{
-  "data": {
-    "type": "post",
-    "id": "1",
-    "attributes": {
-      "message": "From Ember Data"
-    }
-  }
-}
